% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup}
\alias{lookup}
\title{lookup}
\format{
A data frame with (currently) 42,619 rows and 24 columns
}
\source{
\url{https://github.com/drkane/geo-lookups/}
}
\usage{
data(lookup)
}
\description{
DR Kane's lookup of all UK Census LSOAs to higher nesting areas
}
\author{
Francis Barton \email{fbarton@alwaysdata.net}
}
\keyword{datasets}
