% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_slurp.R
\name{statx_slurp}
\alias{statx_slurp}
\title{statx_slurp}
\arguments{
\item{areas_list}{provide a list of area codes for the query}

\item{dataset_name}{name of the dataset on Stat-Xplore (partial/regex works)}

\item{location_level}{return data within an area at this level}

\item{filter_location}{return data within this area. defaults to ".*"}

\item{data_level}{return data at this level}

\item{area_code_lookup}{use this source to lookup area codes at data_level within filter_location}

\item{use_alias}{TRUE by default. Set to FALSE to turn off aliases for location_level and data_level}

\item{batch_size}{If data for more than 1000 area codes are requested then they will be batched into queries of this size. Default is 1000.}

\item{chatty}{TRUE by default. Provides verbose commentary on the query process.}

\item{...}{space to pass parameters to the helper function get_dwp_codes, mainly to do with the number of recent periods (months or quarters) to retrieve data for: provide `periods_tail = n`; see also `periods_head`; you can also tweak the query away from the default of Census geographies to Westminster constituencies, for example, where available, by providing a different value for `geo_type`; you can also change the subset of data from the default by providing a value for `ds`.}
}
\value{
Data frame
}
\description{
statx_slurp
}
\examples{
statx_slurp(
areas_list = "",
dataset_name = "^Carers",
location_level = "lad",
filter_location = "City of London",
data_level = "msoa",
periods_tail = 2,
batch_size = 1000,
use_alias = TRUE,
chatty = TRUE)
}
