% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstatic-package.R
\docType{package}
\name{xstatic-package}
\alias{xstatic}
\alias{xstatic-package}
\title{xstatic: Builds Queries And Retrieves Data From DWP Stat-Xplore}
\description{
Relatively user-friendly approach to formulating queries to the Stat-Xplore API,
    and handling the response.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/francisbarton/xstatic}
}

}
\keyword{internal}
