% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstatic.R
\name{xstatic}
\alias{xstatic}
\title{xstatic}
\usage{
xstatic(
  ben,
  area_codes = NULL,
  filter_level = NULL,
  filter_area = ".*",
  return_level,
  area_code_lookup = NULL,
  geo_level = NULL,
  chatty = TRUE,
  ...
)
}
\arguments{
\item{ben}{name of a benefit dataset on Stat-Xplore (partial/regex works)}

\item{area_codes}{(optional) provide your own vector (list) of area codes for the query (don't use built-in lookup)}

\item{filter_level}{return data within an area at this level}

\item{filter_area}{return data within this area. defaults to ".*" (all)}

\item{return_level}{return data at this level}

\item{area_code_lookup}{use this source to lookup area codes at return_level within filter_area}

\item{geo_level}{If for some reason you need to pass a return_level that is different from the standard options, you will also need to pass a geo_level - an integer between 1 and 7, where 1 means "OA" and 7 refers to "country"}

\item{chatty}{TRUE by default. Provides verbose commentary on the query process.}

\item{...}{space to pass parameters to the helper function get_dwp_codes, mainly to do with the number of recent periods (months or quarters) to retrieve data for: provide `periods_tail = n` (uses 1 - just return most recent period - by default); see also `periods_head`; you can also tweak the query away from the default of Census geographies to Westminster constituencies, for example, where available, by providing a different value for `geo_type`; you can also change the subset of data from the default by providing a different value for `ds`.}
}
\value{
A data frame
}
\description{
xstatic
}
\examples{
xstatic(
ben = "^Carers",
filter_level = "lad", # case-insensitive
filter_area = "City of London", # case sensitive
return_level = "MsOa", # case-insensitive
periods_tail = 2, # starting with most recent available month/quarter
periods_head = 1, # returns penultimate data period only (ie first 1 of 2)
chatty = FALSE)
}
